Turn on I/Q Mode when SoapySDR device is created
The IC-R8600 will not accept any commands via the I/Q USB port
until the I/Q Mode is turned on.  This function was moved from the
stream creation code to the object instantiation.

Also, added Soapy logging to the WinUSBDevice code in preparation
of converting a number of the printf() calls to Soapy log messages.