Add ability to set gains
PRE-AMP, RF, and ATTENUATOR gains may be set.  All follow the Soapy standard of gain in dB
The attenuator is negative dB.
Gain setting was added to the py_test.py script.

According to the Icom manual:
"The gain of the preamplifier is approximately 20 dB on the HF bands,
14dB on the VHF and UHF bands."

It is set to the lower of the two ranges in the driver (14 dB)