Adds antenna Get command
Antenna get is added along with test code to exercise it.
Needed to relocate setting the frequency prior to exercising the antennas
to make sure the radio was tuned in the HF band.  Outside of the HF band,
ANT 1 is always selected, and the radio will respond to antenna Get/Set
commands with an "Invalid Command" error message.

Will need to add this behavior to the list of -TO DO- items for error
checking/handling.