Implements preferred method of global gain setting
The strategy for a single global gain value is:

If the desired gain > 0, turn on the pre-amp (+14dB),
otherwise, turn it off (+0dB)

Then, check to see if any of the attenuator values make sense
(-30dB, -20dB, -10dB, or -0dB)

Finally, make up any remaining difference using the RF Gain
(-63.75dB to 0dB)

getGainRange for overall gain is now valid.

NOTE: This commit partially corrects the issue with Quisk.  Quisk is
still not using the correct gain range (it seems to insist on 0dB to
32dB). At least now if gain is set in the 0dB to 14dB range, this
driver will work.